code:изменение типа переменных

+++ main.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main ()
 {
     cout << "Enter A and B: " << '\n';
-    int a, b, max;
+    double a, b, max;
     cin >> a >> b ;
     cout << "A + B = " << a + b << '\n';
     cout << "A - B = " << a - b << '\n';
